Reducing method scope and renaming test file
Updates to the .gitignore for merlin-data-exchange and merlin-to-hec.  This is a copy from CWMSVUE's .gitignore

Adding MerlinToDss unit test (used for retrieving data from the DB and converting it to DSS), and subsequent changes to support this operation.

Updates to merlin data exchange, storing these for me, but probably won't keep a lot of these changes.

Initial submit for data exchange.

Updating VCS file to point to the right directory

Moving files into a sub-folder called merlin-to-hec to support adding new project.

Updates to the unit test for converting merlin data

Adding logic to the data converter for returning TimeSeriesContainers using HecTime and ZonedDateTime.

Updating the gradle build file, moving over to time series container.  Updating name of project